rds core test suite adjustments
