+++ tests/system-tests/rdscore/internal/rdscorecommon/sriov-validation.go
@@ -9,17 +9,11 @@ import (
 	"github.com/golang/glog"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
-	"github.com/openshift-kni/eco-goinfra/pkg/configmap"
 	"github.com/openshift-kni/eco-goinfra/pkg/deployment"
 	"github.com/openshift-kni/eco-goinfra/pkg/pod"
-	"github.com/openshift-kni/eco-goinfra/pkg/rbac"
 	"github.com/openshift-kni/eco-goinfra/pkg/reportxml"
-	"github.com/openshift-kni/eco-goinfra/pkg/serviceaccount"
-
 	multus "gopkg.in/k8snetworkplumbingwg/multus-cni.v4/pkg/types"
 	corev1 "k8s.io/api/core/v1"
-	rbacv1 "k8s.io/api/rbac/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	. "github.com/openshift-kni/eco-gotests/tests/system-tests/rdscore/internal/rdscoreinittools"
@@ -69,342 +63,6 @@ const (
 	sriovRBACRole3 = "system:openshift:scc:privileged"
 	sriovRBACRole4 = "system:openshift:scc:privileged"
 )
-
-func createServiceAccount(saName, nsName string) {
-	By(fmt.Sprintf("Creating ServiceAccount %q in %q namespace",
-		saName, nsName))
-	glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Creating SA %q in %q namespace",
-		saName, nsName)
-
-	deploySa := serviceaccount.NewBuilder(APIClient, saName, nsName)
-
-	var ctx SpecContext
-
-	deploySa, err := deploySa.Create()
-
-	if err != nil {
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Error creating SA %q in %q namespace: %v",
-			saName, nsName, err)
-	}
-
-	Eventually(func() bool {
-		if !deploySa.Exists() {
-			glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Error creating SA %q in %q namespace",
-				saName, nsName)
-
-			return false
-		}
-
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Created SA %q in %q namespace",
-			deploySa.Definition.Name, deploySa.Definition.Namespace)
-
-		return true
-	}).WithContext(ctx).WithPolling(5*time.Second).WithTimeout(1*time.Minute).Should(BeTrue(),
-		fmt.Sprintf("Failed to create ServiceAccount %q in %q namespace", saName, nsName))
-}
-
-func deleteServiceAccount(saName, nsName string) {
-	By("Removing Service Account")
-	glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Assert SA %q exists in %q namespace",
-		saName, nsName)
-
-	var ctx SpecContext
-
-	if deploySa, err := serviceaccount.Pull(
-		APIClient, saName, nsName); err == nil {
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("ServiceAccount %q found in %q namespace",
-			saName, nsName)
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Deleting ServiceAccount %q in %q namespace",
-			saName, nsName)
-
-		err := deploySa.Delete()
-
-		if err != nil {
-			glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Error deleting ServiceAccount %q in %q namespace: %v",
-				saName, nsName, err)
-		}
-
-		Eventually(func() bool {
-			if deploySa.Exists() {
-				glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Error deleting ServiceAccount %q in %q namespace",
-					saName, nsName)
-
-				return false
-			}
-
-			glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Deleted ServiceAccount %q in %q namespace",
-				saName, nsName)
-
-			return true
-		}).WithContext(ctx).WithPolling(5*time.Second).WithTimeout(1*time.Minute).Should(BeTrue(),
-			fmt.Sprintf("Failed to delete ServiceAccount %q from %q ns", saName, nsName))
-	} else {
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("ServiceAccount %q not found in %q namespace",
-			saName, nsName)
-	}
-}
-
-func deleteClusterRBAC(rbacName string) {
-	By("Deleting Cluster RBAC")
-
-	var ctx SpecContext
-
-	glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Assert ClusterRoleBinding %q exists", rbacName)
-
-	if crbSa, err := rbac.PullClusterRoleBinding(
-		APIClient,
-		rbacName); err == nil {
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("ClusterRoleBinding %q found. Deleting...", rbacName)
-		err := crbSa.Delete()
-
-		if err != nil {
-			glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Error deleting ClusterRoleBinding %q : %v",
-				rbacName, err)
-		}
-
-		Eventually(func() bool {
-			if crbSa.Exists() {
-				glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Error deleting ClusterRoleBinding %q",
-					rbacName)
-
-				return false
-			}
-
-			glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Deleted ClusterRoleBinding %q", rbacName)
-
-			return true
-		}).WithContext(ctx).WithPolling(5*time.Second).WithTimeout(1*time.Minute).Should(BeTrue(),
-			"Failed to delete Cluster RBAC")
-	}
-}
-
-//nolint:unparam
-func createClusterRBAC(rbacName, clusterRole, saName, nsName string) {
-	By("Creating RBAC for SA")
-
-	var ctx SpecContext
-
-	glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Creating ClusterRoleBinding %q", rbacName)
-	crbSa := rbac.NewClusterRoleBindingBuilder(APIClient,
-		rbacName,
-		clusterRole,
-		rbacv1.Subject{
-			Name:      saName,
-			Kind:      "ServiceAccount",
-			Namespace: nsName,
-		})
-
-	crbSa, err := crbSa.Create()
-	if err != nil {
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof(
-			"Error Creating ClusterRoleBinding %q : %v", crbSa.Definition.Name, err)
-	}
-
-	Eventually(func() bool {
-		if !crbSa.Exists() {
-			glog.V(rdscoreparams.RDSCoreLogLevel).Infof(
-				"Error Creating ClusterRoleBinding %q : %v", crbSa.Definition.Name, err)
-
-			return false
-		}
-
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("ClusterRoleBinding %q created:\n\t%v",
-			crbSa.Definition.Name, crbSa)
-
-		return true
-	}).WithContext(ctx).WithPolling(5*time.Second).WithTimeout(1*time.Minute).Should(BeTrue(),
-		"Failed to create ClusterRoleBinding")
-}
-
-func deleteConfigMap(cmName, nsName string) {
-	glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Assert ConfigMap %q exists in %q namespace",
-		cmName, nsName)
-
-	if cmBuilder, err := configmap.Pull(
-		APIClient, cmName, nsName); err == nil {
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("configMap %q found, deleting", cmName)
-
-		var ctx SpecContext
-		err := cmBuilder.Delete()
-		if err != nil {
-			glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Error deleting configMap %q : %v",
-				cmName, err)
-		}
-
-		Eventually(func() bool {
-			if cmBuilder.Exists() {
-				glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Error deleting configMap %q", cmName)
-
-				return false
-			} else {
-				glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Deleted configMap %q in %q namespace",
-					cmName, nsName)
-
-				return true
-			}
-		}).WithContext(ctx).WithPolling(5*time.Second).WithTimeout(1*time.Minute).Should(BeTrue(),
-			"Failed to delete configMap")
-	}
-}
-
-func createConfigMap(cmName, nsName string, data map[string]string) {
-	glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Create ConfigMap %q in %q namespace",
-		cmName, nsName)
-
-	cmBuilder := configmap.NewBuilder(APIClient, cmName, nsName)
-	cmBuilder.WithData(data)
-
-	var ctx SpecContext
-
-	cmResult, err := cmBuilder.Create()
-	if err != nil {
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Error creating ConfigMap %q in %q namespace: %v",
-			cmName, nsName, err)
-	}
-
-	Eventually(func() bool {
-
-		if !cmResult.Exists() {
-			glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Error creating ConfigMap %q in %q namespace",
-				cmName, nsName)
-
-			return false
-		} else {
-			glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Created ConfigMap %q in %q namespace",
-				cmResult.Definition.Name, nsName)
-
-			return true
-		}
-	}).WithContext(ctx).WithPolling(5*time.Second).WithPolling(1*time.Minute).Should(BeTrue(),
-		"Failed to crete configMap")
-}
-
-func deleteDeployments(dName, nsName string) {
-	By(fmt.Sprintf("Removing test deployment %q from %q ns", dName, nsName))
-
-	if deploy, err := deployment.Pull(APIClient, dName, nsName); err == nil {
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Deleting deployment %q from %q namespace",
-			deploy.Definition.Name, nsName)
-
-		err = deploy.DeleteAndWait(300 * time.Second)
-		Expect(err).ToNot(HaveOccurred(),
-			fmt.Sprintf("failed to delete deployment %q", dName))
-	} else {
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Deployment %q not found in %q namespace",
-			dName, nsName)
-	}
-}
-
-func findPodWithSelector(fNamespace, podLabel string) []*pod.Builder {
-	By(fmt.Sprintf("Getting pod(s) matching selector %q", podLabel))
-
-	var (
-		podMatchingSelector []*pod.Builder
-		err                 error
-		ctx                 SpecContext
-	)
-
-	podOneSelector := metav1.ListOptions{
-		LabelSelector: podLabel,
-	}
-
-	Eventually(func() bool {
-		podMatchingSelector, err = pod.List(APIClient, fNamespace, podOneSelector)
-		if err != nil {
-			glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Failed to list pods in %q namespace: %v",
-				fNamespace, err)
-
-			return false
-		}
-
-		if len(podMatchingSelector) == 0 {
-			glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Found 0 pods matching label %q in namespace %q",
-				podLabel, fNamespace)
-
-			return false
-		}
-
-		return true
-	}).WithContext(ctx).WithPolling(15*time.Second).WithTimeout(5*time.Minute).Should(BeTrue(),
-		fmt.Sprintf("Failed to find pod matching label %q in %q namespace", podLabel, fNamespace))
-
-	return podMatchingSelector
-}
-
-func defineContainer(cName, cImage string, cCmd []string, cRequests, cLimits map[string]string) *pod.ContainerBuilder {
-	glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Creating container %q", cName)
-	deployContainer := pod.NewContainerBuilder(cName, cImage, cCmd)
-
-	By("Defining SecurityContext")
-
-	var trueFlag = true
-
-	userUID := new(int64)
-
-	*userUID = 0
-
-	secContext := &corev1.SecurityContext{
-		RunAsUser:  userUID,
-		Privileged: &trueFlag,
-		SeccompProfile: &corev1.SeccompProfile{
-			Type: corev1.SeccompProfileTypeRuntimeDefault,
-		},
-		Capabilities: &corev1.Capabilities{
-			Add: []corev1.Capability{"NET_RAW", "NET_ADMIN", "SYS_ADMIN", "IPC_LOCK"},
-		},
-	}
-
-	By("Setting SecurityContext")
-
-	deployContainer = deployContainer.WithSecurityContext(secContext)
-	glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Container One definition: %#v", deployContainer)
-
-	By("Dropping ALL security capability")
-
-	deployContainer = deployContainer.WithDropSecurityCapabilities([]string{"ALL"}, true)
-
-	By("Adding VolumeMount to container")
-
-	volMount := corev1.VolumeMount{
-		Name:      "configs",
-		MountPath: "/opt/net/",
-		ReadOnly:  false,
-	}
-
-	deployContainer = deployContainer.WithVolumeMount(volMount)
-
-	if len(cRequests) != 0 {
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Processing container's requests")
-
-		containerRequests := corev1.ResourceList{}
-
-		for key, val := range cRequests {
-			glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Parsing container's request: %q - %q", key, val)
-
-			containerRequests[corev1.ResourceName(key)] = resource.MustParse(val)
-		}
-
-		deployContainer = deployContainer.WithCustomResourcesRequests(containerRequests)
-	}
-
-	if len(cLimits) != 0 {
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Processing container's limits")
-
-		containerLimits := corev1.ResourceList{}
-
-		for key, val := range cLimits {
-			glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Parsing container's limit: %q - %q", key, val)
-
-			containerLimits[corev1.ResourceName(key)] = resource.MustParse(val)
-		}
-
-		deployContainer = deployContainer.WithCustomResourcesLimits(containerLimits)
-	}
-
-	glog.V(rdscoreparams.RDSCoreLogLevel).Infof("%q container's  definition:\n%#v", cName, deployContainer)
-
-	return deployContainer
-}
 
 func defineDeployment(containerConfig *corev1.Container, deployName, deployNs, sriovNet, cmName, saName string,
 	deployLabels, nodeSelector map[string]string) *deployment.Builder {
@@ -474,41 +132,6 @@ func defineDeployment(containerConfig *corev1.Container, deployName, deployNs, s
 	}
 
 	return deploy
-}
-
-func verifyMsgInPodLogs(podObj *pod.Builder, msg, cName string, timeSpan time.Time) {
-	glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Parsing duration %q", timeSpan)
-
-	var (
-		podLog string
-		err    error
-		ctx    SpecContext
-	)
-
-	Eventually(func() bool {
-		logStartTimestamp := time.Since(timeSpan)
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("\tTime duration is %s", logStartTimestamp)
-
-		if logStartTimestamp.Abs().Seconds() < 1 {
-			logStartTimestamp, err = time.ParseDuration("1s")
-			Expect(err).ToNot(HaveOccurred(), "Failed to parse time duration")
-		}
-
-		podLog, err = podObj.GetLog(logStartTimestamp, cName)
-
-		if err != nil {
-			glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Failed to get logs from pod %q: %v", podObj.Definition.Name, err)
-
-			return false
-		}
-
-		glog.V(rdscoreparams.RDSCoreLogLevel).Infof("Logs from pod %s:\n%s", podObj.Definition.Name, podLog)
-
-		return true
-	}).WithContext(ctx).WithPolling(5*time.Second).WithTimeout(1*time.Minute).Should(BeTrue(),
-		fmt.Sprintf("Failed to get logs from pod %q", podObj.Definition.Name))
-
-	Expect(podLog).Should(ContainSubstring(msg))
 }
 
 func verifySRIOVConnectivity(nsOneName, nsTwoName, deployOneLabels, deployTwoLabels, targetAddr string) {
